chore(review): purify ui package, reinstate original theme
