+++ packages/extension/src/ui/components/InputText.tsx
@@ -1,4 +1,3 @@
-import { scrollbarStyle } from "@argent-x/ui/src/theme"
 import { isFunction } from "lodash-es"
 import { useId } from "react"
 import { Controller, ControllerProps, FieldValues } from "react-hook-form"
@@ -7,6 +6,7 @@ import TextareaAutosize, {
 } from "react-textarea-autosize"
 import styled, { css } from "styled-components"
 
+import { scrollbarStyle } from "../theme"
 import { isAllowedAddressHexInputValue } from "./utils/isAllowedAddressHexInputValue"
 import { isAllowedNumericInputValue } from "./utils/isAllowedNumericInputValue"
