+++ packages/storybook/src/theme/Colors.stories.mdx
@@ -1,7 +1,5 @@
 import { Meta, ColorPalette, ColorItem } from '@storybook/addon-docs';
 
-import { colors } from "@argent-x/ui/src/theme/theme"
-
 import { groupedColors } from './groupedColors'
 
 <Meta
