+++ packages/ui/package.json
@@ -2,29 +2,21 @@
   "name": "@argent-x/ui",
   "version": "4.8.4",
   "main": "./src/index.ts",
-  "types": "./src/types.d.ts",
   "license": "MIT",
   "devDependencies": {
     "@chakra-ui/cli": "^2.1.8",
-    "@svgr/webpack": "^6.0.0",
-    "@types/lodash-es": "^4.17.6",
     "@types/react": "^18.0.0",
-    "@types/react-copy-to-clipboard": "5.0.4",
     "@types/react-dom": "^18.0.0",
-    "@types/react-measure": "^2.0.8",
-    "@types/styled-components": "^5.1.25",
     "@typescript-eslint/eslint-plugin": "^5.10.1",
     "@typescript-eslint/parser": "^5.10.1",
     "eslint": "^8.7.0",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-webpack-plugin": "^3.1.1",
     "jsdom": "^20.0.0",
-    "typescript": "^4.6.3",
-    "typescript-styled-plugin": "^0.18.2"
+    "typescript": "^4.6.3"
   },
   "scripts": {
-    "gen:theme-typings": "chakra-cli tokens ./src/theme/chakraTheme.ts",
+    "gen:theme-typings": "chakra-cli tokens ./src/theme",
     "lint": "eslint . --cache --ext .ts,.tsx",
     "postinstall": "yarn gen:theme-typings"
   },
@@ -32,27 +24,12 @@
     "@chakra-ui/react": "^2.3.6",
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
-    "@google/model-viewer": "^2.0.0",
-    "@mui/icons-material": "^5.3.1",
-    "@mui/material": "^5.1.0",
-    "@mui/styled-engine-sc": "^5.10.3",
-    "@tippyjs/react": "^4.2.6",
+    "framer-motion": "^7.6.1"
+  },
+  "peerDependencies": {
     "colord": "^2.9.2",
-    "framer-motion": "^7.6.1",
     "lodash-es": "^4.17.21",
-    "qr-code-styling": "^1.6.0-rc.1",
     "react": "^18.0.0",
-    "react-copy-to-clipboard": "^5.0.4",
-    "react-dom": "^18.0.0",
-    "react-dropzone": "^14.0.0",
-    "react-hook-form": "^7.33.0",
-    "react-measure": "^2.5.2",
-    "react-router-dom": "^6.0.1",
-    "react-select": "^5.4.0",
-    "react-textarea-autosize": "^8.3.4",
-    "styled-components": "^5.3.5",
-    "styled-normalize": "^8.0.7",
-    "tracking-link": "^4.0.1",
-    "url-join": "^5.0.0"
+    "react-dom": "^18.0.0"
   }
 }
