--- packages/ui/src/theme/GlobalStyle.tsx
@@ -1,49 +1,0 @@
-import { createGlobalStyle } from "styled-components"
-import { normalize } from "styled-normalize"
-
-export interface GlobalStyleProps {
-  extensionIsInTab: boolean | undefined
-}
-
-export const FixedGlobalStyle = createGlobalStyle<GlobalStyleProps>`
-  ${normalize}
-
-  body {
-    font-family: 'Barlow', sans-serif;
-    -webkit-font-smoothing: antialiased;
-  }
-
-  html, body {
-    min-width: 360px;
-    min-height: 600px;
-
-    width: ${({ extensionIsInTab }) => (extensionIsInTab ? "unset" : "360px")};
-    height: ${({ extensionIsInTab }) => (extensionIsInTab ? "unset" : "600px")};
-    
-    overscroll-behavior: none;
-    -ms-overflow-style: none;  /* IE and Edge */
-    scrollbar-width: none;  /* Firefox */
-    &::-webkit-scrollbar { /* Chrome, Safari, Opera */
-      display: none;
-    }
-  }
-
-  * {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    margin-block: 0;
-  }
-
-  a {
-    text-decoration: none;
-    color: inherit;
-  }
-`
-
-export const ThemedGlobalStyle = createGlobalStyle`
-  body {
-    color: ${({ theme }) => theme.text1};
-    background-color: ${({ theme }) => theme.bg1};
-  }
-`
