+++ packages/ui/src/theme/breakpoints.ts
@@ -1,0 +1,8 @@
+import { pxToRem } from "./utilities/pxToRem"
+
+export const breakpoints = {
+  sm: pxToRem(600),
+  md: pxToRem(900),
+  lg: pxToRem(1200),
+  xl: pxToRem(1536),
+}
