+++ packages/ui/src/theme/index.ts
@@ -1,8 +1,29 @@
-export { FixedGlobalStyle, ThemedGlobalStyle } from "./GlobalStyle"
-export { ThemeProvider } from "./ThemeProvider"
-export { muiTheme } from "./muiTheme"
-export { chakraTheme } from "./chakraTheme"
-export { theme } from "./theme"
-export { scrollbarStyle } from "./scrollbarStyle"
+import { extendTheme } from "@chakra-ui/react"
 
-export type { CustomTheme } from "./theme"
+import { buttonTheme } from "../components/Button"
+import { breakpoints } from "./breakpoints"
+import { colors } from "./colors"
+import { spacing } from "./spacing"
+import { typography } from "./typography"
+
+const extendedTheme = extendTheme({
+  styles: {
+    global: {
+      "html, body": {
+        color: "white",
+        bg: "neutrals.900",
+      },
+    },
+  },
+  breakpoints,
+  ...typography,
+  space: spacing,
+  components: {
+    Button: buttonTheme,
+  },
+})
+
+export const theme = {
+  ...extendedTheme,
+  colors /** omits default chakra colours */,
+}
