+++ packages/ui/src/theme/spacing.ts
@@ -1,0 +1,20 @@
+import { pxToRem } from "./utilities/pxToRem"
+
+export const spacing = {
+  0: pxToRem(0),
+  1: pxToRem(4),
+  2: pxToRem(8),
+  3: pxToRem(12),
+  "3.5": pxToRem(14),
+  4: pxToRem(16),
+  5: pxToRem(20),
+  6: pxToRem(24),
+  7: pxToRem(28),
+  8: pxToRem(32),
+  9: pxToRem(36),
+  10: pxToRem(40),
+  12: pxToRem(48),
+  14: pxToRem(56),
+  16: pxToRem(64),
+  24: pxToRem(96),
+}
