--- packages/ui/src/types.d.ts
@@ -1,6 +1,0 @@
-import { CustomTheme } from "./theme"
-
-declare module "styled-components" {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  export interface DefaultTheme extends CustomTheme {}
-}
